app_streamlit/requirements.txt
streamlit==1.25.0
pandas==1.5.3
psycopg2-binary==2.9.6
plotly==5.15.0
emoji==2.2.0